+++ analyzer.go
@@ -479,27 +479,22 @@ func (analyzer *Analyzer) PageConversions(filter *Filter) (*PageConversionsStats
 func (analyzer *Analyzer) Events(filter *Filter) ([]EventStats, error) {
 	filter = analyzer.getFilter(filter)
 	filter.eventFilter = true
-	outerFilterArgs, outerFilterQuery := filter.query()
-	innerFilterArgs, innerFilterQuery := filter.queryTime()
-	innerFilterArgs = append(innerFilterArgs, outerFilterArgs...)
-	query := fmt.Sprintf(`SELECT event_name,
-		count(DISTINCT visitor_id) visitors,
-		count(1) views,
-		visitors / greatest((
-			SELECT uniq(visitor_id)
-			FROM session
-			WHERE %s
-		), 1) cr,
-		ifNull(toUInt64(avg(nullIf(duration_seconds, 0))), 0) average_duration_seconds,
-		groupUniqArrayArray(event_meta_keys) meta_keys
-		FROM event
-		WHERE %s
-		GROUP BY event_name
-		ORDER BY visitors DESC, event_name
-		%s`, innerFilterQuery, outerFilterQuery, filter.withLimit())
+	args, query := buildQuery(filter, []field{
+		fieldEventName,
+		fieldVisitors,
+		fieldViews,
+		fieldCR,
+		fieldEventTimeSpent,
+		fieldEventMetaKeys,
+	}, []field{
+		fieldEventName,
+	}, []field{
+		fieldVisitors,
+		fieldEventName,
+	})
 	var stats []EventStats
 
-	if err := analyzer.store.Select(&stats, query, innerFilterArgs...); err != nil {
+	if err := analyzer.store.Select(&stats, query, args...); err != nil {
 		return nil, err
 	}
 
@@ -515,31 +510,60 @@ func (analyzer *Analyzer) EventBreakdown(filter *Filter) ([]EventStats, error) {
 		return []EventStats{}, nil
 	}
 
-	outerFilterArgs, outerFilterQuery := filter.query()
-	innerFilterArgs, innerFilterQuery := filter.queryTime()
-	innerFilterArgs = append(innerFilterArgs, filter.EventMetaKey)
-	innerFilterArgs = append(innerFilterArgs, outerFilterArgs...)
-	innerFilterArgs = append(innerFilterArgs, filter.EventMetaKey)
-	query := fmt.Sprintf(`SELECT event_name,
-		count(DISTINCT visitor_id) visitors,
-		count(1) views,
-		visitors / greatest((
-			SELECT uniq(visitor_id)
-			FROM session
-			WHERE %s
-		), 1) cr,
-		ifNull(toUInt64(avg(nullIf(duration_seconds, 0))), 0) average_duration_seconds,
-		event_meta_values[indexOf(event_meta_keys, ?)] meta_value
-		FROM event
-		WHERE %s
-		AND has(event_meta_keys, ?)
-		GROUP BY event_name, meta_value
-		ORDER BY visitors DESC, meta_value
-		%s`, innerFilterQuery, outerFilterQuery, filter.withLimit())
+	args, query := buildQuery(filter, []field{
+		fieldEventName,
+		fieldVisitors,
+		fieldViews,
+		fieldCR,
+		fieldEventTimeSpent,
+		fieldEventMetaValues,
+	}, []field{
+		fieldEventName,
+		fieldEventMetaValues,
+	}, []field{
+		fieldVisitors,
+		fieldEventMetaValues,
+	})
 	var stats []EventStats
 
-	if err := analyzer.store.Select(&stats, query, innerFilterArgs...); err != nil {
+	if err := analyzer.store.Select(&stats, query, args...); err != nil {
 		return nil, err
+	}
+
+	return stats, nil
+}
+
+// EventList returns events as a list. The metadata is grouped as key-value pairs.
+func (analyzer *Analyzer) EventList(filter *Filter) ([]EventListStats, error) {
+	filter = analyzer.getFilter(filter)
+	filter.eventFilter = true
+	args, query := buildQuery(filter, []field{
+		fieldEventName,
+		fieldEventMeta,
+		fieldVisitors,
+		fieldCount,
+	}, []field{
+		fieldEventName,
+		fieldEventMeta,
+	}, []field{
+		fieldCount,
+		fieldEventName,
+	})
+	var stats []EventListStats
+
+	if err := analyzer.store.Select(&stats, query, args...); err != nil {
+		return nil, err
+	}
+
+	// TODO optimize once maps are supported in the driver
+	for i := range stats {
+		stats[i].Meta = make(map[string]string)
+
+		for j := range stats[i].Metadata {
+			key := stats[i].Metadata[j][0].(string)
+			value := stats[i].Metadata[j][1].(string)
+			stats[i].Meta[key] = value
+		}
 	}
 
 	return stats, nil
@@ -587,47 +611,18 @@ func (analyzer *Analyzer) Referrer(filter *Filter) ([]ReferrerStats, error) {
 func (analyzer *Analyzer) Platform(filter *Filter) (*PlatformStats, error) {
 	filter = analyzer.getFilter(filter)
 	table := filter.table()
-	filterArgs, filterQuery := filter.query()
-	args := make([]interface{}, 0, len(filterArgs)*4)
-	var query strings.Builder
+	var args []interface{}
+	query := ""
 
 	if table == "session" {
-		query.WriteString(`SELECT sum(desktop*sign) platform_desktop,
+		filterArgs, filterQuery := filter.query()
+		query = `SELECT sum(desktop*sign) platform_desktop,
 			sum(mobile*sign) platform_mobile,
-			sum(sign)-platform_desktop-platform_mobile platform_unknown, `)
-	} else {
-		args = append(args, filterArgs...)
-		args = append(args, filterArgs...)
-		args = append(args, filterArgs...)
-		query.WriteString(fmt.Sprintf(`SELECT (
-				SELECT uniq(visitor_id)
-				FROM event
-				WHERE %s
-				AND desktop = 1
-				AND mobile = 0
-			) platform_desktop,
-			(
-				SELECT uniq(visitor_id)
-				FROM event
-				WHERE %s
-				AND desktop = 0
-				AND mobile = 1
-			) platform_mobile,
-			(
-				SELECT uniq(visitor_id)
-				FROM event
-				WHERE %s
-				AND desktop = 0
-				AND mobile = 0
-			) platform_unknown, `, filterQuery, filterQuery, filterQuery))
-	}
-
-	query.WriteString(`"platform_desktop" / IF("platform_desktop" + "platform_mobile" + "platform_unknown" = 0, 1, "platform_desktop" + "platform_mobile" + "platform_unknown") AS relative_platform_desktop,
-		"platform_mobile" / IF("platform_desktop" + "platform_mobile" + "platform_unknown" = 0, 1, "platform_desktop" + "platform_mobile" + "platform_unknown") AS relative_platform_mobile,
-		"platform_unknown" / IF("platform_desktop" + "platform_mobile" + "platform_unknown" = 0, 1, "platform_desktop" + "platform_mobile" + "platform_unknown") AS relative_platform_unknown `)
-
-	if table == "session" {
-		query.WriteString(`FROM session s `)
+			sum(sign)-platform_desktop-platform_mobile platform_unknown,
+			"platform_desktop" / IF("platform_desktop" + "platform_mobile" + "platform_unknown" = 0, 1, "platform_desktop" + "platform_mobile" + "platform_unknown") AS relative_platform_desktop,
+			"platform_mobile" / IF("platform_desktop" + "platform_mobile" + "platform_unknown" = 0, 1, "platform_desktop" + "platform_mobile" + "platform_unknown") AS relative_platform_mobile,
+			"platform_unknown" / IF("platform_desktop" + "platform_mobile" + "platform_unknown" = 0, 1, "platform_desktop" + "platform_mobile" + "platform_unknown") AS relative_platform_unknown
+			FROM session s `
 
 		if filter.Path != "" || filter.PathPattern != "" {
 			entryPath, exitPath, eventName := filter.EntryPath, filter.ExitPath, filter.EventName
@@ -635,23 +630,78 @@ func (analyzer *Analyzer) Platform(filter *Filter) (*PlatformStats, error) {
 			innerFilterArgs, innerFilterQuery := filter.query()
 			filter.EntryPath, filter.ExitPath, filter.EventName = entryPath, exitPath, eventName
 			args = append(args, innerFilterArgs...)
-			query.WriteString(fmt.Sprintf(`INNER JOIN (
+			query += fmt.Sprintf(`INNER JOIN (
 				SELECT visitor_id,
 				session_id,
 				path
 				FROM page_view
 				WHERE %s
 			) v
-			ON v.visitor_id = s.visitor_id AND v.session_id = s.session_id `, innerFilterQuery))
+			ON v.visitor_id = s.visitor_id AND v.session_id = s.session_id `, innerFilterQuery)
 		}
 
 		args = append(args, filterArgs...)
-		query.WriteString(fmt.Sprintf(`WHERE %s`, filterQuery))
+		query += fmt.Sprintf(`WHERE %s`, filterQuery)
+	} else {
+		var innerArgs []interface{}
+		innerQuery := ""
+
+		if filter.EntryPath != "" || filter.ExitPath != "" {
+			fields := make([]field, 0, 2)
+
+			if filter.EntryPath != "" {
+				fields = append(fields, fieldEntryPath)
+			}
+
+			if filter.ExitPath != "" {
+				fields = append(fields, fieldExitPath)
+			}
+
+			innerArgs, innerQuery = joinSessions(filter, table, fields)
+			filter.EntryPath, filter.ExitPath = "", ""
+		}
+
+		filterArgs, filterQuery := filter.query()
+		args = make([]interface{}, 0, len(filterArgs)*3+len(innerArgs)*3)
+		args = append(args, innerArgs...)
+		args = append(args, filterArgs...)
+		args = append(args, innerArgs...)
+		args = append(args, filterArgs...)
+		args = append(args, innerArgs...)
+		args = append(args, filterArgs...)
+		query = fmt.Sprintf(`SELECT (
+				SELECT uniq(visitor_id)
+				FROM event v
+				%s
+				WHERE %s
+				AND desktop = 1
+				AND mobile = 0
+			) platform_desktop,
+			(
+				SELECT uniq(visitor_id)
+				FROM event v
+				%s
+				WHERE %s
+				AND desktop = 0
+				AND mobile = 1
+			) platform_mobile,
+			(
+				SELECT uniq(visitor_id)
+				FROM event v
+				%s
+				WHERE %s
+				AND desktop = 0
+				AND mobile = 0
+			) platform_unknown,
+			"platform_desktop" / IF("platform_desktop" + "platform_mobile" + "platform_unknown" = 0, 1, "platform_desktop" + "platform_mobile" + "platform_unknown") AS relative_platform_desktop,
+			"platform_mobile" / IF("platform_desktop" + "platform_mobile" + "platform_unknown" = 0, 1, "platform_desktop" + "platform_mobile" + "platform_unknown") AS relative_platform_mobile,
+			"platform_unknown" / IF("platform_desktop" + "platform_mobile" + "platform_unknown" = 0, 1, "platform_desktop" + "platform_mobile" + "platform_unknown") AS relative_platform_unknown `,
+			innerQuery, filterQuery, innerQuery, filterQuery, innerQuery, filterQuery)
 	}
 
 	stats := new(PlatformStats)
 
-	if err := analyzer.store.Get(stats, query.String(), args...); err != nil {
+	if err := analyzer.store.Get(stats, query, args...); err != nil {
 		return nil, err
 	}