+++ filter_test.go
@@ -81,14 +81,19 @@ func TestFilter_QueryFields(t *testing.T) {
 	filter.BrowserVersion = "89"
 	filter.Platform = PlatformUnknown
 	filter.ScreenClass = "XXL"
+	filter.ScreenWidth = "1920"
+	filter.ScreenHeight = "1080"
 	filter.UTMSource = "source"
 	filter.UTMMedium = "medium"
 	filter.UTMCampaign = "campaign"
 	filter.UTMContent = "content"
 	filter.UTMTerm = "term"
+	filter.EventMeta = map[string]string{
+		"foo": "bar",
+	}
 	filter.validate()
 	args, query := filter.queryFields()
-	assert.Len(t, args, 18)
+	assert.Len(t, args, 21)
 	assert.Equal(t, "/", args[0])
 	assert.Equal(t, "/entry", args[1])
 	assert.Equal(t, "/exit", args[2])
@@ -102,17 +107,20 @@ func TestFilter_QueryFields(t *testing.T) {
 	assert.Equal(t, BrowserEdge, args[10])
 	assert.Equal(t, "89", args[11])
 	assert.Equal(t, "XXL", args[12])
-	assert.Equal(t, "source", args[13])
-	assert.Equal(t, "medium", args[14])
-	assert.Equal(t, "campaign", args[15])
-	assert.Equal(t, "content", args[16])
-	assert.Equal(t, "term", args[17])
-	assert.Equal(t, "path = ? AND entry_path = ? AND exit_path = ? AND language = ? AND country_code = ? AND city = ? AND referrer = ? AND referrer_name = ? AND os = ? AND os_version = ? AND browser = ? AND browser_version = ? AND screen_class = ? AND utm_source = ? AND utm_medium = ? AND utm_campaign = ? AND utm_content = ? AND utm_term = ? AND desktop = 0 AND mobile = 0 ", query)
+	assert.Equal(t, uint16(1920), args[13])
+	assert.Equal(t, uint16(1080), args[14])
+	assert.Equal(t, "source", args[15])
+	assert.Equal(t, "medium", args[16])
+	assert.Equal(t, "campaign", args[17])
+	assert.Equal(t, "content", args[18])
+	assert.Equal(t, "term", args[19])
+	assert.Equal(t, "bar", args[20])
+	assert.Equal(t, "path = ? AND entry_path = ? AND exit_path = ? AND language = ? AND country_code = ? AND city = ? AND referrer = ? AND referrer_name = ? AND os = ? AND os_version = ? AND browser = ? AND browser_version = ? AND screen_class = ? AND screen_width = ? AND screen_height = ? AND utm_source = ? AND utm_medium = ? AND utm_campaign = ? AND utm_content = ? AND utm_term = ? AND desktop = 0 AND mobile = 0 AND event_meta_values[indexOf(event_meta_keys, 'foo')] = ? ", query)
 	filter.EventName = "event"
 	args, query = filter.queryFields()
-	assert.Len(t, args, 17)
-	assert.Equal(t, "event", args[16])
-	assert.Equal(t, "path = ? AND language = ? AND country_code = ? AND city = ? AND referrer = ? AND referrer_name = ? AND os = ? AND os_version = ? AND browser = ? AND browser_version = ? AND screen_class = ? AND utm_source = ? AND utm_medium = ? AND utm_campaign = ? AND utm_content = ? AND utm_term = ? AND event_name = ? AND desktop = 0 AND mobile = 0 ", query)
+	assert.Len(t, args, 22)
+	assert.Equal(t, "event", args[20])
+	assert.Equal(t, "path = ? AND entry_path = ? AND exit_path = ? AND language = ? AND country_code = ? AND city = ? AND referrer = ? AND referrer_name = ? AND os = ? AND os_version = ? AND browser = ? AND browser_version = ? AND screen_class = ? AND screen_width = ? AND screen_height = ? AND utm_source = ? AND utm_medium = ? AND utm_campaign = ? AND utm_content = ? AND utm_term = ? AND event_name = ? AND desktop = 0 AND mobile = 0 AND event_meta_values[indexOf(event_meta_keys, 'foo')] = ? ", query)
 }
 
 func TestFilter_QueryFieldsInvert(t *testing.T) {
@@ -132,14 +140,19 @@ func TestFilter_QueryFieldsInvert(t *testing.T) {
 	filter.BrowserVersion = "!89"
 	filter.Platform = "!" + PlatformUnknown
 	filter.ScreenClass = "!XXL"
+	filter.ScreenWidth = "!1920"
+	filter.ScreenHeight = "!1080"
 	filter.UTMSource = "!source"
 	filter.UTMMedium = "!medium"
 	filter.UTMCampaign = "!campaign"
 	filter.UTMContent = "!content"
 	filter.UTMTerm = "!term"
+	filter.EventMeta = map[string]string{
+		"foo": "!bar",
+	}
 	filter.validate()
 	args, query := filter.queryFields()
-	assert.Len(t, args, 18)
+	assert.Len(t, args, 21)
 	assert.Equal(t, "/", args[0])
 	assert.Equal(t, "/entry", args[1])
 	assert.Equal(t, "/exit", args[2])
@@ -153,17 +166,20 @@ func TestFilter_QueryFieldsInvert(t *testing.T) {
 	assert.Equal(t, BrowserEdge, args[10])
 	assert.Equal(t, "89", args[11])
 	assert.Equal(t, "XXL", args[12])
-	assert.Equal(t, "source", args[13])
-	assert.Equal(t, "medium", args[14])
-	assert.Equal(t, "campaign", args[15])
-	assert.Equal(t, "content", args[16])
-	assert.Equal(t, "term", args[17])
-	assert.Equal(t, "path != ? AND entry_path != ? AND exit_path != ? AND language != ? AND country_code != ? AND city != ? AND referrer != ? AND referrer_name != ? AND os != ? AND os_version != ? AND browser != ? AND browser_version != ? AND screen_class != ? AND utm_source != ? AND utm_medium != ? AND utm_campaign != ? AND utm_content != ? AND utm_term != ? AND (desktop = 1 OR mobile = 1) ", query)
+	assert.Equal(t, uint16(1920), args[13])
+	assert.Equal(t, uint16(1080), args[14])
+	assert.Equal(t, "source", args[15])
+	assert.Equal(t, "medium", args[16])
+	assert.Equal(t, "campaign", args[17])
+	assert.Equal(t, "content", args[18])
+	assert.Equal(t, "term", args[19])
+	assert.Equal(t, "bar", args[20])
+	assert.Equal(t, "path != ? AND entry_path != ? AND exit_path != ? AND language != ? AND country_code != ? AND city != ? AND referrer != ? AND referrer_name != ? AND os != ? AND os_version != ? AND browser != ? AND browser_version != ? AND screen_class != ? AND screen_width != ? AND screen_height != ? AND utm_source != ? AND utm_medium != ? AND utm_campaign != ? AND utm_content != ? AND utm_term != ? AND (desktop = 1 OR mobile = 1) AND event_meta_values[indexOf(event_meta_keys, 'foo')] != ? ", query)
 	filter.EventName = "!event"
 	args, query = filter.queryFields()
-	assert.Len(t, args, 17)
-	assert.Equal(t, "event", args[16])
-	assert.Equal(t, "path != ? AND language != ? AND country_code != ? AND city != ? AND referrer != ? AND referrer_name != ? AND os != ? AND os_version != ? AND browser != ? AND browser_version != ? AND screen_class != ? AND utm_source != ? AND utm_medium != ? AND utm_campaign != ? AND utm_content != ? AND utm_term != ? AND event_name != ? AND (desktop = 1 OR mobile = 1) ", query)
+	assert.Len(t, args, 22)
+	assert.Equal(t, "event", args[20])
+	assert.Equal(t, "path != ? AND entry_path != ? AND exit_path != ? AND language != ? AND country_code != ? AND city != ? AND referrer != ? AND referrer_name != ? AND os != ? AND os_version != ? AND browser != ? AND browser_version != ? AND screen_class != ? AND screen_width != ? AND screen_height != ? AND utm_source != ? AND utm_medium != ? AND utm_campaign != ? AND utm_content != ? AND utm_term != ? AND event_name != ? AND (desktop = 1 OR mobile = 1) AND event_meta_values[indexOf(event_meta_keys, 'foo')] != ? ", query)
 }
 
 func TestFilter_QueryFieldsNull(t *testing.T) {
@@ -183,30 +199,35 @@ func TestFilter_QueryFieldsNull(t *testing.T) {
 	filter.BrowserVersion = "null"
 	filter.Platform = "null"
 	filter.ScreenClass = "null"
+	filter.ScreenWidth = "null"
+	filter.ScreenHeight = "Null"
 	filter.UTMSource = "null"
 	filter.UTMMedium = "null"
 	filter.UTMCampaign = "null"
 	filter.UTMContent = "null"
 	filter.UTMTerm = "null"
+	filter.EventMeta = map[string]string{
+		"foo": "null",
+	}
 	filter.validate()
 	args, query := filter.queryFields()
-	assert.Len(t, args, 18)
+	assert.Len(t, args, 21)
 
 	for i := 0; i < len(args); i++ {
 		assert.Empty(t, args[i])
 	}
 
-	assert.Equal(t, "path = ? AND entry_path = ? AND exit_path = ? AND language = ? AND country_code = ? AND city = ? AND referrer = ? AND referrer_name = ? AND os = ? AND os_version = ? AND browser = ? AND browser_version = ? AND screen_class = ? AND utm_source = ? AND utm_medium = ? AND utm_campaign = ? AND utm_content = ? AND utm_term = ? AND desktop = 0 AND mobile = 0 ", query)
+	assert.Equal(t, "path = ? AND entry_path = ? AND exit_path = ? AND language = ? AND country_code = ? AND city = ? AND referrer = ? AND referrer_name = ? AND os = ? AND os_version = ? AND browser = ? AND browser_version = ? AND screen_class = ? AND screen_width = ? AND screen_height = ? AND utm_source = ? AND utm_medium = ? AND utm_campaign = ? AND utm_content = ? AND utm_term = ? AND desktop = 0 AND mobile = 0 AND event_meta_values[indexOf(event_meta_keys, 'foo')] = ? ", query)
 	filter.EventName = "null"
 	filter.validate()
 	args, query = filter.queryFields()
-	assert.Len(t, args, 17)
+	assert.Len(t, args, 22)
 
 	for i := 0; i < len(args); i++ {
 		assert.Empty(t, args[i])
 	}
 
-	assert.Equal(t, "path = ? AND language = ? AND country_code = ? AND city = ? AND referrer = ? AND referrer_name = ? AND os = ? AND os_version = ? AND browser = ? AND browser_version = ? AND screen_class = ? AND utm_source = ? AND utm_medium = ? AND utm_campaign = ? AND utm_content = ? AND utm_term = ? AND event_name = ? AND desktop = 0 AND mobile = 0 ", query)
+	assert.Equal(t, "path = ? AND entry_path = ? AND exit_path = ? AND language = ? AND country_code = ? AND city = ? AND referrer = ? AND referrer_name = ? AND os = ? AND os_version = ? AND browser = ? AND browser_version = ? AND screen_class = ? AND screen_width = ? AND screen_height = ? AND utm_source = ? AND utm_medium = ? AND utm_campaign = ? AND utm_content = ? AND utm_term = ? AND event_name = ? AND desktop = 0 AND mobile = 0 AND event_meta_values[indexOf(event_meta_keys, 'foo')] = ? ", query)
 }
 
 func TestFilter_QueryFieldsPlatform(t *testing.T) {
@@ -364,7 +385,10 @@ func TestFilter_Fields(t *testing.T) {
 
 	filter.validate()
 	filter.EventName = "event"
-	assert.Equal(t, "path,language,country_code,city,referrer,referrer_name,os,os_version,browser,browser_version,screen_class,utm_source,utm_medium,utm_campaign,utm_content,utm_term,event_name,desktop,mobile", filter.fields())
+	filter.EventMeta = map[string]string{
+		"foo": "bar",
+	}
+	assert.Equal(t, "path,entry_path,exit_path,language,country_code,city,referrer,referrer_name,os,os_version,browser,browser_version,screen_class,utm_source,utm_medium,utm_campaign,utm_content,utm_term,event_name,desktop,mobile,event_meta_keys,event_meta_values", filter.fields())
 }
 
 func pastDay(n int) time.Time {